refactor: move sapling note to the correct module
