+++ lib/uniffi-zcash/src/udl/zcash_primitives/sapling/note/note.udl
