+++ lib/uniffi-zcash/src/zcash_primitives/sapling/mod.rs
@@ -12,3 +12,6 @@ pub use self::sapling_ivk::*;
 
 mod value;
 pub use self::value::*;
+
+mod note;
+pub use self::note::*;
