+++ lib/uniffi-zcash/src/zcash_primitives/sapling/note/mod.rs
@@ -1,0 +1,2 @@
+mod note;
+pub use self::note::*;
